model_zoo/official/cv/unet/requirements.txt
opencv-python >= 4.1.2.30
pillow >= 6.2.0
